+++ entity/Player.java
@@ -13,17 +13,23 @@ public class Player extends Entity {
     GamePanel gp;
     KeyHandler keyH;
 
+    public final int screenX, screenY;
+
     public Player(GamePanel gp, KeyHandler keyH) {
         this.gp = gp;
         this.keyH = keyH;
+
+        screenX = gp.screenWidth / 2 - (gp.tileSize / 2);
+        screenY = gp.screenHeight / 2 - (gp.tileSize / 2);
 
         setDefaultValues();
         getPlayerImage();
     }
 
     public void setDefaultValues() {
-        x = 100;
-        y = 100;
+        /*posicion del personaje */
+        worldX = gp.tileSize * 20;
+        worldY = gp.tileSize * 5;
         speed = 3;
         direction = "idle";
 
@@ -53,22 +59,24 @@ public class Player extends Entity {
     }
 
     public void update() {
-        if(keyH.upPressed || keyH.downPressed || keyH.leftPressed || keyH.rightPressed){
-            /*Aca tendria que meter todos los if de abajo si no tubiese los sprites "Idle" */
+        if (keyH.upPressed || keyH.downPressed || keyH.leftPressed || keyH.rightPressed) {
+            /*
+             * Aca tendria que meter todos los if de abajo si no tubiese los sprites "Idle"
+             */
         }
 
         if (keyH.upPressed) {
             direction = "up";
-            y -= speed;
+            worldY -= speed;
         } else if (keyH.downPressed) {
             direction = "down";
-            y += speed;
+            worldY += speed;
         } else if (keyH.leftPressed) {
             direction = "left";
-            x -= speed;
+            worldX -= speed;
         } else if (keyH.rightPressed) {
             direction = "right";
-            x += speed;
+            worldX += speed;
         } else {
             direction = "idle";
         }
@@ -140,8 +148,8 @@ public class Player extends Entity {
                 }
                 break;
 
-            }
-            spriteCounter++;
+        }
+        spriteCounter++;
 
         if (spriteCounter > 60) {
             if (spriteNum == 1) {
@@ -153,6 +161,6 @@ public class Player extends Entity {
             }
             spriteCounter = 0;
         }
-        g2.drawImage(image, x, y, gp.tileSize, gp.tileSize, null);
+        g2.drawImage(image, screenX, screenY, gp.tileSize, gp.tileSize, null);
     }
 }